// src/app/core/models/User.ts
export interface User {
  beanName: string;
  createdAt: string;
  createdBy: string;
  createdByUid: number;
  firstName: string;
  id: number;
  lastName: string;
  majorSpecificType: string;
  middleName: string;
  modifiedAt: string;
  modifiedBy: string;
  modifiedByUid: number;
  serviceUser: boolean;
  specificUser: boolean;
  sponsoredUser: boolean;
  titleAfter: string;
  titleBefore: string;
}